refactor: 기존 버전을 V1 패키지로 이동, 컨트롤러에 V2를 주입
