+++ product-service/src/main/java/com/musinsam/productservice/application/service/V1/ProductServiceApiV1.java
@@ -1,4 +1,4 @@
-package com.musinsam.productservice.application.service;
+package com.musinsam.productservice.application.service.V1;
 
 import com.musinsam.common.user.CurrentUserDtoApiV1;
 import com.musinsam.productservice.application.dto.request.ReqProductPatchByProductIdDtoApiV1;
